+++ tfimm/utils/__init__.py
@@ -7,5 +7,5 @@ from .cache import (  # noqa: F401
     set_model_cache,
 )
 from .constants import *  # noqa: F401
-from .etc import to_2tuple  # noqa: F401
+from .etc import make_divisible, to_2tuple  # noqa: F401
 from .timm import load_pth_url_weights, load_timm_weights  # noqa: F401
